+++ shell.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	char *line;
+	char *line, **str;
 
 	while (1)
 	{
@@ -18,21 +18,14 @@ int main(void)
 			write(STDOUT_FILENO, "$ ", 2);
 		}
 		line = read_line();
-		if (line == NULL)
-		{
-			if (isatty(STDIN_FILENO))
-			{
-				write(STDOUT_FILENO, "\n", 1);
-				break;
-			}
-			
-		}
-		if (_strlen(line) == 0 || white_space(line))
+		if (*line == '\n')
 		{
 			free(line);
 			continue;
 		}
-		execute_command(line);
+
+		str = split_string(line);
+		child_process(str, environ);
 		free(line);
 	}
 	return (0);
